Jameson
yes ok i want that and to make the defulat unexspanded file editor open in the middle and be a rectangel like rectangular horazontaly. also make to boarder be a darker color than the editor and anagin i want only one exit button on the editor and the text and the exspander button on the left side. no questions get to updating!